cleanup_all script completed

+++ tools/cleanup_all-vagrant_init_destroy_boxes.sh
@@ -4,11 +4,16 @@ set -o pipefail
 
 TMPDIR=${TMPDIR:-/var/tmp/vagrant_init_destroy_boxes}
 
-for DIR in `find /var/tmp -maxdepth 1 -mindepth 1 -type d`; do
-  VAGRANT_BOX_NAME=$(basename $DIR)
-  echo "*** $DIR"
+for DIR in `find $TMPDIR -maxdepth 1 -mindepth 1 -type d`; do
+  VAGRANT_BOX_NAME=$(awk -F\" '/config.vm.box =/ { print $2 }' $DIR/Vagrantfile)
+  VAGRANT_BOX_PROVIDER=${VAGRANT_BOX_NAME##*-}
+
+  echo "*** $DIR | $VAGRANT_BOX_NAME | $VAGRANT_BOX_PROVIDER"
   cd $DIR
   vagrant destroy -f
-  virsh --connect=qemu:///system vol-delete --pool default --vol ${VAGRANT_BOX_NAME}_vagrant_box_image_0.img
+  if [ "$VAGRANT_BOX_PROVIDER" = "libvirt" ]; then
+    virsh --connect=qemu:///system vol-delete --pool default --vol ${VAGRANT_BOX_NAME}_vagrant_box_image_0.img
+  fi
+  cd -
   rm -rf $DIR
 done
